Don't print bad edict lines during preparse

+++ libs/edict.rb
@@ -200,17 +200,23 @@ private
 
   EdictMarshal = Struct.new(:e,:k)
 
-  def Edict.preparse
+  def Edict.load_marshaled
+    if File.exist?($DICT_DIR+"/edict/edict.marshal") &&
+       File.stat($DICT_DIR+"/edict/edict.marshal").mtime > File.stat($DICT_DIR+'/edict/edict').mtime
+      edict_marshal = File.open($DICT_DIR+"/edict/edict.marshal", "rb") {|f| Marshal.load(f)}
+      return [edict_marshal.e, edict_marshal.k]
+    end
+
     e = Hash.new {|hh,kk| hh[kk] = []}
     k = Hash.new {|hh,kk| hh[kk] = []}
 
     print "preparsing... "
-
     lines = Utf8.readlines($DICT_DIR+'/edict/edict','euc-jp')
 
     # save a copy of edict as utf-8 purely for convenience
     File.open($DICT_DIR+'/edict/edict.utf8','w') {|f| lines.each {|line| f.puts line}}
 
+    bad_lines = []
     lines[1..-1].each_with_index do |line,i|
       if m = line.match(/(.+?) \[(.+?)\] (\/.+\/)/)
         e[m[1]] << line
@@ -219,22 +225,24 @@ private
         e[m[1]] << line
         k[m[1]] << line
       else
-        puts "skipping bad edict line (#{i+1}): `#{line}`"
+        bad_lines << line
       end
     end
 
+    print "(#{bad_lines.size} bad lines)... " if !bad_lines.empty?
+
+    print "marshaling... "
     e.default = nil
     k.default = nil
     File.open($DICT_DIR+"/edict/edict.marshal", "wb") {|f| Marshal.dump(EdictMarshal.new(e,k), f)}
+
+    [e, k]
   end
 
   def Edict.load!
     print "Loading Edict... "
     Progress.new do |pr|
-      preparse unless File.exist?($DICT_DIR+"/edict/edict.marshal")
-      edict_marshal = File.open($DICT_DIR+"/edict/edict.marshal", "rb") {|f| Marshal.load(f)}
-      @e = edict_marshal.e
-      @k = edict_marshal.k
+      @e, @k = load_marshaled
     end
   end
